requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.1
openpyxl==3.1.5
requests==2.32.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
plotly==5.24.1
st_pages==1.0.1
scikit-learn==1.5.2
category_encoders==2.6.4
imbalanced-learn==0.12.3
streamlit_elements==0.1.0
kaggle==1.6.17
openai==0.28.0
chardet==5.2.0
streamlit_cookies_manager==0.2.0